requirements.txt
appdirs==1.4.4
astroid==2.4.2
atomicwrites==1.4.0
autopep8 @ file:///tmp/build/80754af9/autopep8_1620866417880/work
backcall==0.2.0
black==21.9b0
Bottleneck @ file:///C:/ci/bottleneck_1607557040328/work
Brotli==1.0.9
Cartopy @ file:///C:/ci/cartopy_1613152141732/work
certifi==2021.5.30
click==8.0.1
colorama==0.4.3
cycler==0.10.0
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.0.7
distlib==0.3.1
filelock==3.0.12
flake8==3.8.4
Flask==2.0.1
Flask-Compress==1.10.1
fonttools==4.25.0
iniconfig==1.1.1
ipykernel==5.5.4
ipython==7.23.1
ipython-genutils==0.2.0
isort==5.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.2
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mccabe==0.6.1
mkl-fft==1.3.0
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
munkres==1.1.4
mypy-extensions==0.4.3
numexpr @ file:///C:/ci/numexpr_1618856728739/work
numpy==1.21.2
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
packaging==20.7
pandas @ file:///C:/ci/pandas_1629682571109/work
parso==0.8.2
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.3.0
plotly @ file:///tmp/build/80754af9/plotly_1626182254919/work
pluggy==0.13.1
prompt-toolkit==3.0.18
py==1.9.0
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1615748559966/work
pyflakes==2.2.0
Pygments==2.9.0
pylint==2.6.0
pyparsing==2.4.7
pyserial==3.5
pyshp @ file:///tmp/build/80754af9/pyshp_1610641829514/work
PySimpleGUI==4.47.0
pytest==6.1.2
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pywin32==300
pyzmq==22.0.3
regex==2021.8.28
scipy @ file:///C:/ci/scipy_1630606917240/work
Shapely==1.7.0
six==1.15.0
tenacity @ file:///C:/ci/tenacity_1626248381338/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli==1.2.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.2
virtualenv==20.0.34
wcwidth==0.2.5
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
